Move mutation to confirm screen
