+++ client/packages/app/features/rider/rider-screen.tsx
@@ -1,11 +1,7 @@
-import { AuthContext } from '@my/app/provider/AuthProvider'
 import { useRouterStore, useStore, useUserStore } from '@my/app/store'
-import { TOAST_DURATION } from '@my/app/util'
-import getServerUrl from '@my/app/util/getServerUrl'
-import { Button, Paragraph, Separator, YStack, useToastController } from '@my/ui'
+import { Button, Paragraph, Separator, YStack } from '@my/ui'
 import { ChevronLeft } from '@tamagui/lucide-icons'
-import { useMutation } from '@tanstack/react-query'
-import { useContext, useState } from 'react'
+import { useState } from 'react'
 import { LngLat } from 'react-map-gl'
 import { useRouter } from 'solito/navigation'
 import { MapBox } from '../common/MapBox'
@@ -14,54 +10,13 @@ import { RiderConfirm } from './rider-confirm'
 import { RiderSearch } from './rider-search'
 
 export function RiderHomeScreen() {
-  const auth = useContext(AuthContext)
-
   const router = useRouter()
-  const toast = useToastController()
-  const SERVER_URL = getServerUrl()
   const [pickUplngLat, setPickUpLnglat] = useState<LngLat | undefined>()
   const [destLngLat, setDestLnglat] = useState<LngLat | undefined>()
   const [page, setPage] = useState(2)
 
   const store = useStore(useUserStore, (store) => store)
   const routeStore = useStore(useRouterStore, (store) => store)
-
-  const { mutate, isPending, isIdle } = useMutation({
-    mutationFn: async () =>
-      fetch(`${SERVER_URL}/rider/trip`, {
-        method: 'POST',
-        referrer: SERVER_URL,
-        body: JSON.stringify({
-          id: store?.id,
-          time: routeStore?.pickupTime?.toISOString(),
-          pickup: {
-            lng: routeStore?.pickup?.longitude.toPrecision(17),
-            lat: routeStore?.pickup?.latitude.toPrecision(17),
-          },
-          destination: {
-            lng: routeStore?.destination?.longitude.toPrecision(17),
-            lat: routeStore?.destination?.latitude.toPrecision(17),
-          },
-        }),
-        headers: {
-          Authorization: auth!.session!.access_token,
-        },
-      }),
-    onError(err, v, c) {
-      console.error(err)
-      toast.show('Issue while requesting your ride', {
-        message: err?.message,
-        duration: TOAST_DURATION,
-      })
-    },
-    async onSuccess(data, v, c) {
-      var res = await data.json()
-      toast.show(res.message, {
-        message: res.requestId,
-        duration: TOAST_DURATION,
-      })
-    },
-  })
 
   return (
     <YStack
